either camelcase acronyms or don't, but pick one
